+++ lib/src/queue_item.dart
@@ -1,4 +1,6 @@
-class QueueItem<T> {
+import 'package:equatable/equatable.dart';
+
+class QueueItem<T> extends Equatable {
   final T value;
   final int priority;
   final String? key;
@@ -7,5 +9,11 @@ class QueueItem<T> {
     required this.value,
     int? priority,
     this.key,
-  }) : priority = priority ?? DateTime.now().millisecondsSinceEpoch;
+  }) : priority = priority ?? DateTime.now().microsecondsSinceEpoch;
+
+  @override
+  List<Object?> get props => [
+        priority,
+        key,
+      ];
 }